backend/foodgram/requirements.txt
django==4.1
djangorestframework==3.14.0
requests==2.28.2
Pillow==9.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
djoser==2.1.0
django-filter==23.1
django-extra-fields==3.0.2